requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
tenacity==8.2.3
python-dotenv==1.0.0
email-validator==2.1.0
requests==2.31.0
google-auth-oauthlib==1.2.0
google-auth==2.25.2
httpx==0.25.1
